requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
celery==5.3.0
cffi==2.0.0
channels==4.2.2
channels_redis==4.2.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.1
daphne==4.2.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
# Caching dependencies
django-crispy-forms==2.1
django-debug-toolbar==6.0.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
gunicorn==23.0.0
hiredis==3.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pymemcache==4.0.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.2.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
zope.interface==8.0
